fix example perspective projection field of view
Unlike the old `Matrix44.perspective_projection`, `glm.perspective`
expects its field of view argument in radians, not degrees as were
implied by the magnitude of 45.

Fix by using Python's built in radians conversion function. 45 radians
corresponds (wrapped) to 58.31 degrees so the field of view is specified
as 60 degrees to produce a negligible change in example output.