add ability to attach TextureArray as layered framebuffer
Allows the use of `gl_Layer` in a geometry shader to direct a particular
primitive to be rendered onto a particular layer of a texture array.

Attempting this previously raised an error so there are few backwards
compatibility concerns. Like other types of framebuffers, the mipmap
level is not exposed to Python and is always 0.

Tested and used to good effect in my program. This seems to be supported
in OpenGL 3.3 as well.

See https://www.khronos.org/opengl/wiki/Framebuffer_Object#Layered_Images
and https://www.khronos.org/opengl/wiki/Geometry_Shader#Layered_rendering .